% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.pf1.R
\name{default.pf}
\alias{default.pf}
\title{Data adaptive candidate vector of penalty factors for L1 penalty in conditional logistic regression}
\usage{
default.pf(
  X,
  Y,
  stratum,
  nfolds = 10,
  alpha = 1,
  standardize = TRUE,
  type.step1,
  p
)
}
\arguments{
\item{X}{A matrix of covariates, with the number of rows equaling the number
of observations.}

\item{Y}{A binary response variable.}

\item{stratum}{A numeric vector with stratum membership of each observation.}

\item{nfolds}{The number of folds used in cross-validation. Default is 10.}

\item{alpha}{The elastic net mixing parameter, a number between 0 and 1.
alpha=0 would give pure ridge; alpha=1 gives lasso. Pure ridge penalty is never obtained in this implementation since alpha must be positive.}

\item{standardize}{Should the covariates be standardized, a logical value.}

\item{type.step1}{Should the tentative model be fit on all covariates jointly or to each block separately.}

\item{p}{The sizes of blocks of covariates,
a numerical vector of the length equal to the number of blocks,
and with the sum equal to the number of penalized covariates.
If missing, all covariates are treated the same and a single penalty is applied.}
}
\value{
A numeric vector of length 1 to 3 (depending on the problem)
giving L1 penalties
}
\description{
Performs cross validation to determine reasonable default values for L1 penalty
in a conditional logistic regression
}
