% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.clr.R
\name{stable.clr}
\alias{stable.clr}
\title{Stability selection based on penalized conditional logistic regression}
\usage{
stable.clr(
  response,
  stratum,
  penalized,
  unpenalized = NULL,
  lambda.seq = NULL,
  alpha = 1,
  B = 100,
  parallel = TRUE,
  standardize = FALSE,
  event
)
}
\arguments{
\item{response}{The response variable, either a 0/1 vector or a factor with two levels.}

\item{stratum}{A numeric vector with stratum membership of each observation.}

\item{penalized}{A matrix of penalized covariates.}

\item{unpenalized}{A matrix of additional unpenalized covariates.}

\item{lambda.seq}{a sequence of non-negative values to be used as tuning
parameters for L1}

\item{alpha}{The elastic net mixing parameter, a number between 0 and 1.
alpha=0 would give pure ridge; alpha=1 gives lasso. Pure ridge penalty is never obtained in this implementation since alpha must be positive.}

\item{B}{A single positive number for the number of subsamples.}

\item{parallel}{Logical. Should the computation be parallelized?}

\item{standardize}{Should the covariates be standardized, a logical value.}

\item{event}{If response is a factor, the level that
should be considered a success in the logistic regression.}
}
\value{
A numeric vector \code{Pilambda}
giving selection probabilities for each penalized covariate.
}
\description{
Performs stability selection for conditional logistic regression models with L1 and/or L2 penalty.
}
